refactor: consolidate error classes into @markuplint/shared
Move all cross-package error classes (ParserError, TargetParserError,
ConfigParserError, ConfigLoadError, InvalidSelectorError,
UnexpectedCallError) and isFatalError() guard into
@markuplint/shared/src/errors/. Original packages re-export for
backward compatibility.
