+++ packages/@markuplint/file-resolver/src/config-load-error.ts
@@ -1,10 +1,2 @@
-export class ConfigLoadError extends Error {
-	filePath: string;
-	name = 'ConfigLoadError';
-	referrer: string;
-	constructor(message: string, filePath: string, referrer: string) {
-		super(message + ` in ${referrer}`);
-		this.filePath = filePath;
-		this.referrer = referrer;
-	}
-}
+// Re-export from @markuplint/shared — this class is now defined there.
+export { ConfigLoadError } from '@markuplint/shared';
