+++ packages/@markuplint/ml-core/src/ml-dom/node/unexpected-call-error.ts
@@ -1,1 +1,2 @@
-export class UnexpectedCallError extends Error {}
+// Re-export from @markuplint/shared — this class is now defined there.
+export { UnexpectedCallError } from '@markuplint/shared';
