+++ packages/@markuplint/shared/src/errors/index.ts
@@ -1,0 +1,6 @@
+export { ConfigLoadError } from './config-error.js';
+export { isFatalError } from './guards.js';
+export { ConfigParserError, ParserError, TargetParserError } from './parser-error.js';
+export type { ParserErrorInfo } from './parser-error.js';
+export { InvalidSelectorError } from './selector-error.js';
+export { UnexpectedCallError } from './unexpected-call-error.js';
