+++ packages/@markuplint/shared/src/errors/unexpected-call-error.ts
@@ -1,0 +1,1 @@
+export class UnexpectedCallError extends Error {}
