+++ packages/@markuplint/shared/src/index.ts
@@ -1,2 +1,3 @@
+export * from './errors/index.js';
 export * from './functions.js';
 export type * from './types.js';
